External-app : Send a reply card when receiving a card (#2259)
