--- src/test/clientApp/src/main/resources/application.yml
@@ -1,6 +1,0 @@
-spring:
-  main:
-    web-application-type: none
-logging:
-  level:
-    org.springframework.web.client.RestTemplate: DEBUG
