+++ src/test/externalApp/externalApp.gradle
@@ -43,9 +43,14 @@ docker {
 
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation "org.opfab:cards-client-data:${project.version}"
+
 	compileOnly 'org.projectlombok:lombok'
 	annotationProcessor 'org.projectlombok:lombok'
+	implementation project(':tools:generic:utilities'), project(':tools:spring:spring-utilities')
 	testImplementation('org.springframework.boot:spring-boot-starter-test') {
+
+	implementation project(':client:cards-client-data')
 	}
 }
