--- src/test/externalApp/src/main/java/org/opfab/externalApp/service/ExternalAppServiceImpl.java
@@ -1,36 +1,0 @@
-/* Copyright (c) 2018-2021, RTE (http://www.rte-france.com)
- * See AUTHORS.txt
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- * SPDX-License-Identifier: MPL-2.0
- * This file is part of the OperatorFabric project.
- */
-
-package org.opfab.externalApp.service;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-@Slf4j
-public class ExternalAppServiceImpl implements ExternalAppService {
-
-	@Override
-	public void receiveCard(Optional<JsonNode> requestBody) {
-		log.info("card reception from Card Publication Service {} : \n\n", requestBody);
-	}
-
-	@Override
-	public void deleteCard(String id) {
-		log.info("Card suppression from Card Publication Service cardId = {} : \n\n", id);
-	}
-
-	public String WelcomeMessage() {
-		return   "Welcome to External Application";
-	}
-
-}
