+++ src/test/externalApp/src/main/java/org/opfab/externalapp/security/AuthToken.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021, RTE (http://www.rte-france.com)
+/* Copyright (c) 2022, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,7 +7,7 @@
  * This file is part of the OperatorFabric project.
  */
 
-package org.opfab.client.security;
+package org.opfab.externalapp.security;
 
 import lombok.Getter;
 
@@ -16,11 +16,4 @@ public class AuthToken {
 
     @Getter
     private String access_token;
-    private long expires_in;
-    private long refresh_expires_in;
-    private String refresh_token;
-    private String token_type;
-    private String id_token;
-    private String scope;
-    private String session_state;
 }
